+++ engines/zvision/render_table.h
@@ -66,7 +66,7 @@ public:
 
 	const Common::Point convertWarpedCoordToFlatCoord(const Common::Point &point);
 
-	void mutateImage(uint16 *sourceBuffer, uint16* destBuffer, uint16 destWidth);
+	void mutateImage(uint16 *sourceBuffer, uint16* destBuffer, uint32 destWidth);
 	void generateRenderTable();
 
 	void setPanoramaFoV(float fov);
